UPDATE: route handlers for GET, POST  api requests - removed POST req handler for '/:dummyId' FILE RENAME: - dummyRoute.js > dummyRouter.js
